Moved NodeViewWrapper to its own module
