+++ components/common/CharmEditor/CharmEditor.tsx
@@ -247,7 +247,7 @@ const StyledReactBangleEditor = styled(ReactBangleEditor)`
     color: ${({ theme }) => theme.palette.code.color};
     display: inline-block;
     font-size: 85%;
-    height: 100%;
+    height: fit-content;
     tab-size: 4;
     caret-color: black;
   }
